Refactor billing.py into a module
Tested via `deployer generate-cost-table  --output=terminal`
